+++ config/test.exs
@@ -1,0 +1,3 @@
+import Config
+
+config :save_it, :start_telegram_bot, false